Phase 9: Syscall.java の sys_* シグネチャを long に昇格
68 個の sys_* メソッドのシグネチャを変更:
  int sys_X( int bx, int cx, int dx, int si, int di )
  → long sys_X( long bx, long cx, long dx, long si, long di )

これにより SyscallAmd64 から呼ぶ際の (int)a1 キャストによる高位
スタックアドレスの切り詰めが不要になる。amd64 の dispatch から
24 箇所の (int) キャストを除去した。

副次的変更:
  - Syscall.call() の戻り値も int → long
  - SyscallI386.call() の signature と ret 型を long に変更
  - Cpu.java (i386) で reg[AX] = (int)syscall.call(...) と明示キャスト
  - sys_setuid/setgid/umask 内で (int)bx 明示キャスト
  - sys_mmap/sys_mremap で arg() ヘルパー呼び出し時に (int)bx を保存
  - sys_writev で内部 ret/len を long に
  - SyscallAmd64 のコメントと dispatch 表記を更新

回帰テスト: 42 本中 41 PASS / 0 FAIL / 1 SKIP (sys_chmod64 のみ環境制約)。
シグナル配信や fork/exec を含む全テストが引き続き動作することを確認。

+++ src/main/java/emulin/Cpu.java
@@ -452,7 +452,7 @@ public class Cpu extends AbstractCpu
 
   // INT命令
   void interrupt( ) {
-    reg[ AX ] = syscall.call( reg[ AX ], reg[BX], reg[CX], reg[DX], reg[SI], reg[DI] );
+    reg[ AX ] = (int)syscall.call( reg[ AX ], reg[BX], reg[CX], reg[DX], reg[SI], reg[DI] );
   }
 
   // POP命令
